SO-4025: correct license
