+++ snomed/com.b2international.snowowl.snomed.datastore.tests/src/com/b2international/snowowl/snomed/datastore/compare/ReferenceSetCompareTest.java
@@ -1,6 +1,18 @@
-/*******************************************************************************
- * Copyright (c) 2020 B2i Healthcare. All rights reserved.
- *******************************************************************************/
+/*
+ * Copyright 2012 B2i Healthcare Pte Ltd, http://b2i.sg
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.b2international.snowowl.snomed.datastore.compare;
 
 import static org.junit.Assert.assertEquals;
@@ -11,7 +23,7 @@ import java.util.Map;
 import org.junit.Test;
 
 import com.b2international.commons.exceptions.BadRequestException;
-import com.b2international.snowowl.core.compare.CompareSetResult;
+import com.b2international.snowowl.core.compare.ConceptCompareSetResult;
 import com.b2international.snowowl.snomed.common.SnomedRf2Headers;
 import com.b2international.snowowl.snomed.common.SnomedTerminologyComponentConstants;
 import com.b2international.snowowl.snomed.core.domain.SnomedConcept;
@@ -63,7 +75,7 @@ public class ReferenceSetCompareTest {
 		rf1.setMembers(new SnomedReferenceSetMembers(ImmutableList.of(member1), null, 1, 1));
 		rf2.setMembers(new SnomedReferenceSetMembers(ImmutableList.of(member2), null, 1, 1));
 		
-		CompareSetResult<SnomedReferenceSetMember> result = comparator.doCompare(rf1, rf2);
+		ConceptCompareSetResult<SnomedReferenceSetMember> result = comparator.doCompare(rf1, rf2);
 		assertEquals(0, result.getRemovedMembers().size());
 		assertEquals(0, result.getAddedMembers().size());
 		assertEquals(0, result.getChangedMembers().size());
@@ -145,7 +157,7 @@ public class ReferenceSetCompareTest {
 		rf1.setMembers(new SnomedReferenceSetMembers(ImmutableList.of(equalMember1, removedMember), null, 2, 2));
 		rf2.setMembers(new SnomedReferenceSetMembers(ImmutableList.of(equalMember2, changedMember, addedMember), null, 3, 3));
 		
-		CompareSetResult<SnomedReferenceSetMember> result = comparator.doCompare(rf1, rf2);
+		ConceptCompareSetResult<SnomedReferenceSetMember> result = comparator.doCompare(rf1, rf2);
 		assertEquals(1, result.getRemovedMembers().size());
 		assertEquals(1, result.getAddedMembers().size());
 		assertEquals(1, result.getChangedMembers().size());
@@ -212,7 +224,7 @@ public class ReferenceSetCompareTest {
 		rf1.setMembers(new SnomedReferenceSetMembers(ImmutableList.of(equalMember1, removedMember), null, 2, 2));
 		rf2.setMembers(new SnomedReferenceSetMembers(ImmutableList.of(equalMember2, changedMember, addedMember), null, 3, 3));
 		
-		CompareSetResult<SnomedReferenceSetMember> result = comparator.doCompare(rf1, rf2);
+		ConceptCompareSetResult<SnomedReferenceSetMember> result = comparator.doCompare(rf1, rf2);
 		assertEquals(1, result.getRemovedMembers().size());
 		assertEquals(1, result.getAddedMembers().size());
 		assertEquals(1, result.getChangedMembers().size());
